Script for calculating sensitivity
This file calculated sensitivity by taking the difference between average anpp and anpp in a given year divided by the difference between average precipitation and precipitation in a given year. I think there may be issues with this method as it is currently calculated. I have included comments discussing this in the script.